# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go Copilot API
  description: |
    Authentication and Role-Based Access Control (RBAC) API.

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    Sensitive endpoints are rate-limited. When limits are exceeded, a 429 response is returned with a Retry-After header.

    ## Permissions
    Access to resources is controlled via RBAC. Users are assigned roles, which contain permissions.
    Permission format: `resource:action` (e.g., `users:read`, `roles:create`)

    ## Common Flow Examples

    ### 1. User Registration and Login
    ```bash
    # Register a new user
    curl -X POST http://localhost:8080/api/v1/auth/register \
      -H "Content-Type: application/json" \
      -d '{"email": "user@example.com", "password": "SecurePass123!", "full_name": "John Doe"}'

    # Login to get tokens
    curl -X POST http://localhost:8080/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"email": "user@example.com", "password": "SecurePass123!"}'
    ```

    ### 2. Making Authenticated Requests
    ```bash
    # Get current user profile
    curl -X GET http://localhost:8080/api/v1/auth/me \
      -H "Authorization: Bearer <access_token>"

    # List users (requires users:list permission)
    curl -X GET http://localhost:8080/api/v1/users \
...
    ```

    ### 3. Refreshing Tokens
...

...

...

...
servers:
...
paths:
...
  /roles:
    get:
      tags:
        - Roles
      summary: List roles
      description: Get all roles
      operationId: listRoles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires roles:list permission
    post:
      tags:
        - Roles
      summary: Create role
      description: Create a new role
      operationId: createRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires roles:create permission
        '409':
          description: Role name already exists
...
  /permissions:
    get:
      tags:
        - Permissions
      summary: List permissions
      description: Get all permissions, optionally filtered by resource
      operationId: listPermissions
      security:
        - bearerAuth: []
      parameters:
        - name: resource
          in: query
          schema:
            type: string
          description: Filter by resource (e.g., "users", "roles")
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires permissions:list permission
    post:
      tags:
        - Permissions
      summary: Create permission
      description: Create a custom permission
      operationId: createPermission
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermissionRequest'
      responses:
        '201':
          description: Permission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires permissions:create permission
        '409':
          description: Permission code already exists
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  schemas:
...
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        is_system:
          type: boolean
        is_default:
          type: boolean
        priority:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateRoleRequest:
      type: object
      required:
        - name
        - display_name
      properties:
        name:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
          example: content_editor
        display_name:
          type: string
          example: Content Editor
        description:
          type: string
        permission_ids:
          type: array
          items:
            type: string
            format: uuid
...
    PermissionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        resource:
          type: string
          example: users
        action:
          type: string
          example: read
        code:
          type: string
          example: users:read
        description:
          type: string
        is_system:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePermissionRequest:
      type: object
      required:
        - resource
        - action
      properties:
        resource:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
          example: articles
        action:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
          example: publish
        description:
          type: string
          example: Publish articles
...